refactor: move stream types to handler packages, fix systemd write permissions
- Move stream type constants from shared ipc package to respective handlers
  (filebrowser, dbus) for better encapsulation
- Move PkgUpdateProgress type and writer to dbus package
- Remove unused OperationCanceller (operations.go)
- Consolidate RegisterTerminalHandlers into RegisterAllHandlers
- Add ReadWritePaths to systemd service for apt config, systemd timers,
  and wireguard configs (fixes write failures with ProtectSystem=strict)
